Refactor AuthBackgroundView
